--- src/env.validation.ts
@@ -1,43 +1,0 @@
-import { plainToInstance } from 'class-transformer'
-import {
-    IsEnum,
-    IsNumber,
-    IsString,
-    validateSync,
-    IsOptional,
-    IsNotEmpty,
-} from 'class-validator'
-
-enum Environment {
-    Development = 'development',
-    Production = 'production',
-    Test = 'test',
-}
-
-export class EnvironmentVariables {
-    @IsEnum(Environment)
-    NODE_ENV!: Environment
-
-    @IsNumber()
-    @IsOptional()
-    PORT?: number
-
-    @IsString()
-    @IsNotEmpty()
-    DATABASE_URL!: string
-}
-
-export function validate(config: Record<string, unknown>) {
-    const validatedConfig = plainToInstance(EnvironmentVariables, config, {
-        enableImplicitConversion: true,
-    })
-
-    const errors = validateSync(validatedConfig, {
-        skipMissingProperties: false,
-    })
-
-    if (errors.length > 0) {
-        throw new Error(errors.toString())
-    }
-    return validatedConfig
-}
